+++ packages/react/src/evm/providers/thirdweb-provider.tsx
@@ -30,7 +30,7 @@ interface ThirdwebProviderProps<TChains extends Chain[]>
    * />
    * ```
    */
-  supportedWallets?: WalletConfig<any, any>[];
+  supportedWallets?: WalletConfig<any>[];
 }
 
 /**